+++ src/main/java/Knowledge/InstantaneousKnowledge.java
@@ -9,19 +9,19 @@ import java.util.List;
 
 public class InstantaneousKnowledge extends Knowledge
 {
-  public final int askPrice;
-  public final int bidPrice;
+  public final long askPrice;
+  public final long bidPrice;
   public final List<SellOffer> sellOffers;
   public final List<BuyOffer> buyOffers;
   public final Assets freeAssets;
 
   public InstantaneousKnowledge(
-      Agent holder,
-      int askPrice,
-      int bidPrice,
-      List<SellOffer> sellOffers,
-      List<BuyOffer> buyOffers,
-      Assets freeAssets)
+          Agent holder,
+          long askPrice,
+          long bidPrice,
+          List<SellOffer> sellOffers,
+          List<BuyOffer> buyOffers,
+          Assets freeAssets)
   {
     super(holder);
     this.askPrice = askPrice;
