+++ src/main/java/Simulation/Transaction.java
@@ -2,8 +2,8 @@ package Simulation;
 
 public class Transaction {
     public final Agent buyer, seller;
-    public final int cashQuantity, stockQuantity, stockPrice;
-    public Transaction (Agent buyer, Agent seller, int cashQuantity, int stockQuantity, int stockPrice) {
+    public final long cashQuantity, stockQuantity, stockPrice;
+    public Transaction (Agent buyer, Agent seller, long cashQuantity, long stockQuantity, long stockPrice) {
         this.buyer = buyer;
         this.seller = seller;
         this.cashQuantity = cashQuantity;
